requirements.txt
psycopg2
psaw
praw
sqlalchemy
pyyaml